refactoring to get rid of ChatVideo
